Styled the UI using Material Design components and Guidelines besides adding a few icons
