+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        applicationId "com.example.tipcalculator"
+        applicationId "com.example.tiptime"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
@@ -16,6 +16,9 @@ android {
         buildFeatures{
             viewBinding = true
         }
+
+        vectorDrawables.useSupportLibrary = true
+
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
