+++ app/src/main/res/values/themes.xml
@@ -12,5 +12,12 @@
         <!-- Status bar color. -->
         <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
         <!-- Customize your theme here. -->
+        <!--Text Input Fields-->
+        <item name="textInputStyle">@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox</item>
+        <!--For Radio Buttons-->
+        <item name="radioButtonStyle">@style/Widget.TipTime.CompoundButton.RadioButton</item>
+        <!-- For Switches -->
+        <item name="switchStyle">@style/Widget.TipTime.CompoundButton.Switch</item>
+
     </style>
 </resources>
