+++ app/src/test/java/com/example/tiptime/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.tipcalculator
+package com.example.tiptime
 
 import org.junit.Test
